Move vs Build
The main function now calculates max damage, min damage, min moves to KO, and max moves to KO, against all of the build options more likely to occur than 1%, for each pokemon, for each move. This is a huge step towards being complete. It's going to need a ton of optimization.